chore: renamed ServiceClient into LtiServiceClient (same for interface) feat: added more contextual params to UserAuthenticatorInterface chore: fixed doc and tests
