+++ doc/message/platform-originating-messages.md
@@ -222,16 +222,22 @@ You have to specify how to provide platform authentication, for example:
 ```php
 <?php
 
+use OAT\Library\Lti1p3Core\Registration\RegistrationInterface;
 use OAT\Library\Lti1p3Core\Security\User\UserAuthenticatorInterface;
-use OAT\Library\Lti1p3Core\Security\User\UserAuthenticationResult;
-use OAT\Library\Lti1p3Core\Security\User\UserAuthenticationResultInterface;
+use OAT\Library\Lti1p3Core\Security\User\Result\UserAuthenticationResult;
+use OAT\Library\Lti1p3Core\Security\User\Result\UserAuthenticationResultInterface;
 use OAT\Library\Lti1p3Core\User\UserIdentity;
+use Psr\Http\Message\ServerRequestInterface;
 
 $userAuthenticator = new class implements UserAuthenticatorInterface
 {
-   public function authenticate(string $loginHint): UserAuthenticationResultInterface
-   {
-       // Perform user authentication based on the login hint (ex: owned session, LDAP, external auth service, etc)
+   public function authenticate(
+       RegistrationInterface $registration,
+       ServerRequestInterface $request,
+       string $loginHint
+   ): UserAuthenticationResultInterface {
+       // Perform user authentication based on the registration, request or login hint
+       // (ex: owned session, LDAP, external auth service, etc)
        ...       
 
         return new UserAuthenticationResult(
