+++ src/Security/Oidc/OidcAuthenticator.php
@@ -95,6 +95,8 @@ class OidcAuthenticator
             }
 
             $authenticationResult = $this->authenticator->authenticate(
+                $registration,
+                $request,
                 $oidcRequest->getParameters()->getMandatory('login_hint')
             );
