+++ tests/Traits/SecurityTestingTrait.php
@@ -35,10 +35,11 @@ use OAT\Library\Lti1p3Core\Security\Key\KeyInterface;
 use OAT\Library\Lti1p3Core\Security\Nonce\Nonce;
 use OAT\Library\Lti1p3Core\Security\Nonce\NonceInterface;
 use OAT\Library\Lti1p3Core\Security\Nonce\NonceRepositoryInterface;
-use OAT\Library\Lti1p3Core\Security\User\UserAuthenticationResult;
-use OAT\Library\Lti1p3Core\Security\User\UserAuthenticationResultInterface;
+use OAT\Library\Lti1p3Core\Security\User\Result\UserAuthenticationResult;
+use OAT\Library\Lti1p3Core\Security\User\Result\UserAuthenticationResultInterface;
 use OAT\Library\Lti1p3Core\Security\User\UserAuthenticatorInterface;
 use OAT\Library\Lti1p3Core\Util\Generator\IdGeneratorInterface;
+use Psr\Http\Message\ServerRequestInterface;
 
 trait SecurityTestingTrait
 {
@@ -152,8 +153,11 @@ trait SecurityTestingTrait
                 $this->withAnonymous = $withAnonymous;
             }
 
-            public function authenticate(string $loginHint): UserAuthenticationResultInterface
-            {
+            public function authenticate(
+                RegistrationInterface $registration,
+                ServerRequestInterface $request,
+                string $loginHint
+            ): UserAuthenticationResultInterface {
                 return new UserAuthenticationResult(
                     $this->withAuthenticationSuccess,
                     $this->withAnonymous ? null : $this->createTestUserIdentity()
